changed to all doubles

+++ lib/polarToCartesian.hpp
@@ -4,13 +4,13 @@
 namespace ngPlasmid
 {
     Point polarToCartesian(
-        float centerX,
-        float centerY,
-        float radius,
-        float angleInDegrees
+        double centerX,
+        double centerY,
+        double radius,
+        double angleInDegrees
     ) {
         Point res;
-        float angleInRadians = (angleInDegrees - 90) * M_PI / 180.0;
+        double angleInRadians = (angleInDegrees - 90) * M_PI / 180.0;
         res.x = centerX + (radius * std::cos(angleInRadians));
         res.y = centerY + (radius * std::sin(angleInRadians));
         return res;
